requirements.txt
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cloudinary==1.44.1
crispy-bootstrap5==2025.6
cryptography==46.0.3
Django==5.2.8
django-background-tasks==1.2.8
django-cloudinary-storage==0.3.0
django-crispy-forms==2.5
django-social-share==2.3.0
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.11
mysql-connector-python==9.5.0
packaging==25.0
pillow==12.0.0
pycparser==2.23
PyMySQL==1.1.2
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
